refactor make crossvalsets file
